requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.19
faiss-cpu==1.9.0
sentence-transformers==3.4.1
torch==2.5.1
pymupdf==1.25.5
pdfplumber==0.11.4
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
html2text==2024.2.26
rank-bm25==0.2.2
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
streamlit==1.41.1
openai==1.68.2
groq==0.13.1
langdetect==1.0.9
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.67.1
loguru==0.7.3
numpy==1.26.4
pandas==2.2.3
nltk==3.9.1
pytest==8.3.4
pytest-asyncio==0.24.0